requirements.txt
djangorestframework
django
Pillow
django-cors-headers
